bitbake: cooker: Use a queue to feed parsing jobs
Curerntly, recipes to parse are split into equal groups and passed to
each parse thread at the start of parsing. We can replace this with
a queue and collect a new job as each parsing process becomes idle
to better spread load in the case of slow parsing jobs.

Some of the data we need has to be passed in at fork time since it
can't be pickled, so the job to parse is only referenced as an index
in that list.

This should better spread load for slow to parse recipes such as those
with many class extensions.

(Bitbake rev: 12325a0ad77c71afa3a42d2b97783e9d4456595c)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>